refactor: post 관련 상태 전역 상태 관리로 리팩토링
