+++ src/entities/post/index.ts
@@ -1,0 +1,2 @@
+export type { Post } from "./model/types"
+export { usePost } from "./model/usePost"
