+++ src/entities/post/model/types.ts
@@ -1,0 +1,11 @@
+import { User } from "@/entities/user"
+
+export type Post = {
+  id: number
+  title: string
+  body?: string
+  userId?: number
+  tags?: string[]
+  reactions?: { likes: number; dislikes: number }
+  author?: User
+}
