+++ src/entities/user/model/types.ts
@@ -1,0 +1,1 @@
+export type User = { id: number; username: string; image?: string }
